refactor: move Column to root of ir mod
